Lots has changed. Individual permutations are stored as bits in a long, so all permutations for a group are stored in a Long[]. With this new storage format both filtering and generation consistent pattern have changed to match this.
Logic has been added to delay permutation generation until a threshold of solved cells is reached to reduce the number of stored permutations.

Both in-line filtering and early pruning happens during generation to reduce the number of permutations stored and stop recursively generation a permutation early it's partial state already doesn't fit the partial solution.

Timing class has been added to provide useful time data on solves as well as a couple other metric.

Also a GenerationStatWriter class used to help determine efficiency around the pre-allocation that needs to happen as just using an Array to store the permutations.